requirements.txt
flask
python-dotenv
faiss-cpu
PyPDF2
tqdm
openai
requests
huggingface-hub
sentence-transformers
